feature_store/requirements.txt
redis
xgboost
numpy
scikit-learn
matplotlib
pandas
feast[gcp, redis]
google-cloud-bigquery
google-cloud-storage
